Deprecate `Youtubedl-no-compression` header
`Accept-Encoding`: identity does the same thing.

Backported from networking rework (#2861)

Authored by: coletdjnz
