+++ yt_dlp/downloader/external.py
@@ -529,7 +529,9 @@ class FFmpegFD(ExternalFD):
         selected_formats = info_dict.get('requested_formats') or [info_dict]
         for i, fmt in enumerate(selected_formats):
             if fmt.get('http_headers') and re.match(r'^https?://', fmt['url']):
-                headers_dict = handle_youtubedl_headers(fmt['http_headers'])
+                if 'Youtubedl-no-compression' in fmt['http_headers']:  # compat
+                    fmt['http_headers'].pop('Youtubedl-no-compression', None)
+                    fmt['http_headers']['Accept-encoding'] = 'identity'
                 # Trailing \r\n after each HTTP header is important to prevent warning from ffmpeg/avconv:
                 # [http @ 00000000003d2fa0] No trailing CRLF found in HTTP header.
                 args.extend(['-headers', ''.join(f'{key}: {val}\r\n' for key, val in headers_dict.items())])
