+++ yt_dlp/utils.py
@@ -48,6 +48,7 @@ import xml.etree.ElementTree
 import zlib
 
 from .compat import functools  # isort: split
+from .compat.utils import *  # noqa: F401, F403
 from .compat import (
     compat_etree_fromstring,
     compat_expanduser,
@@ -1351,25 +1352,12 @@ def _create_http_connection(ydl_handler, http_class, is_https, *args, **kwargs):
     return hc
 
 
-def handle_youtubedl_headers(headers):
-    filtered_headers = headers
-
-    if 'Youtubedl-no-compression' in filtered_headers:
-        filtered_headers = {k: v for k, v in filtered_headers.items() if k.lower() != 'accept-encoding'}
-        del filtered_headers['Youtubedl-no-compression']
-
-    return filtered_headers
-
-
 class YoutubeDLHandler(urllib.request.HTTPHandler):
     """Handler for HTTP requests and responses.
 
     This class, when installed with an OpenerDirector, automatically adds
     the standard headers to every HTTP request and handles gzipped and
-    deflated responses from web servers. If compression is to be avoided in
-    a particular request, the original request in the program code only has
-    to include the HTTP header "Youtubedl-no-compression", which will be
-    removed before making the real request.
+    deflated responses from web servers.
 
     Part of this code was copied from:
 
@@ -1435,7 +1423,9 @@ class YoutubeDLHandler(urllib.request.HTTPHandler):
         if 'Accept-encoding' not in req.headers:
             req.add_header('Accept-encoding', ', '.join(SUPPORTED_ENCODINGS))
 
-        req.headers = handle_youtubedl_headers(req.headers)
+        if 'Youtubedl-no-compression' in req.headers:  # compat
+            req.headers.pop('Youtubedl-no-compression', None)
+            req.headers['Accept-encoding'] = 'identity'
 
         return super().do_request_(req)
 
@@ -6506,15 +6496,3 @@ class FormatSorter:
                 format['abr'] = format.get('tbr') - format.get('vbr', 0)
 
         return tuple(self._calculate_field_preference(format, field) for field in self._order)
-
-
-# Deprecated
-has_certifi = bool(certifi)
-has_websockets = bool(websockets)
-
-
-def load_plugins(name, suffix, namespace):
-    from .plugins import load_plugins
-    ret = load_plugins(name, suffix)
-    namespace.update(ret)
-    return ret
